major update
Adding example to be used, ECommerce Cart and interaction between the user and the system.

Knows Issues:
When a client add more than 1 products the quantity value is increaced, But when the client delete all the products from the cart and try add the same one again, the quantity is not restored to 0/1

This will be fixed in the next Release.