Shopping Cart Feature
Adding disabled/enabled methods to avoid client missunderstanding and proceed only to checkout page once we have items in the cart.

If not, the client will not be able to proceed.

As well, I as a clien cant click on "vaciar carrito" if we doesnt have items in the cart. To avoid problems during the event genrator.