Improvements
new items:
+checkout.html

updated:
*aplicado.html
*style.scss
*db.,js
*products.js
*style.css

Adding easy way to increace or decreace the item quantity directly before of move to checkout page

We also update the information in the local storage to retrieve the information.

Know Issues:
1.- Filter: When the user clicks on Buscar we are not able to keep the text there. Can be a problem for the client in case to want to double check the word to search.

As well the problem with add/remove class was solved with a walk around, using Swal as alert to just avoid a possible increace.

As well issue related to UX was solved, the client during the confirmation Screen, have the way to increace the item number to more than we have in the stock.

Checkout page was added to retrieve the information from shopping cart.

Methods will be implemented to show the product list and display the information during the checkout process